TINY-8391: Move relevant table tests from core to model
