+++ modules/tinymce/src/models/dom/test/ts/browser/table/NewCellRowEventsTest.ts
@@ -4,9 +4,9 @@ import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 
-import * as TableTestUtils from '../../module/test/TableUtils';
+import * as TableTestUtils from '../../module/table/TableTestUtils';
 
-describe('browser.tinymce.core.table.NewCellRowEventsTest', () => {
+describe('browser.tinymce.models.dom.table.NewCellRowEventsTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     indent: false,
     valid_styles: {
