+++ modules/tinymce/src/models/dom/test/ts/browser/table/TableSizingModeTest.ts
@@ -6,7 +6,7 @@ import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 
-describe('browser.tinymce.core.table.TableSizingModeTest', () => {
+describe('browser.tinymce.models.dom.table.TableSizingModeTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     width: 800,
     base_url: '/project/tinymce/js/tinymce'
